+++ app/Http/Controllers/Api/V1_0/VolunteerEducationController.php
@@ -8,7 +8,7 @@ use App\Http\Requests\Api\V1_0\UpdateEducationRequest;
 use Gate;
 use App\Http\Controllers\Api\BaseAuthController;
 use App\Exceptions\AccessDeniedException;
-use App\Transformers\VolunteerEducationTransformer;
+use App\Transformers\Volunteer\VolunteerEducationTransformer;
 use App\Education;
 use App\Services\JwtService;
 use App\Services\TransformerService;
@@ -39,7 +39,7 @@ class VolunteerEducationController extends BaseAuthController
 
         $manager = TransformerService::getManager();
         $resource = TransformerService::getResourceCollection($educations,
-            'App\Transformers\VolunteerEducationTransformer', 'educations');
+            'App\Transformers\Volunteer\VolunteerEducationTransformer', 'educations');
 
         return response()->json($manager->createData($resource)->toArray(), 200);
     }
